FIX errorHandle midleware

+++ src/models/mysql/articles.js
@@ -9,51 +9,75 @@ import { ClientError } from '../../utils/errors.js';
 export async function getAllModel({ category }) {
 	const connection = await connectionDB();
 
-	if (!category) {
-		const [articles] = await connection.execute(
-			'SELECT a.article_id, a.title, a.content, c.category FROM articles a LEFT JOIN categories c ON a.article_id = c.article_id;',
+	try {
+		if (!category) {
+			const [articles] = await connection.execute(
+				'SELECT a.article_id, a.title, a.content, c.category FROM articles a LEFT JOIN categories c ON a.article_id = c.article_id;',
+			);
+
+			return { code: 200, response: articles };
+		}
+
+		const [article] = await connection.execute(
+			'SELECT a.article_id, a.title, a.content, c.category FROM articles a LEFT JOIN categories c ON a.article_id = c.article_id WHERE category = ?;',
+			[category],
 		);
 
-		return articles;
+		if (article.length === 0) throw new ClientError('category not found');
+
+		return { code: 200, response: article };
+	} catch (error) {
+		if (error.statusCode === 400)
+			throw {
+				code: error.statusCode,
+				response: error.message,
+			};
+
+		throw { completeError: error };
+	} finally {
+		connection.end();
 	}
-
-	const [article] = await connection.execute(
-		'SELECT a.article_id, a.title, a.content, c.category FROM articles a LEFT JOIN categories c ON a.article_id = c.article_id WHERE category = ?;',
-		[category],
-	);
-
-	if (article.length === 0) throw new ClientError('category not found');
-
-	return article;
 }
 
 export async function getByIdModel({ id }) {
 	const connection = await connectionDB();
 
-	const [article] = await connection.execute(
-		'SELECT a.article_id, a.title, a.content, c.category FROM articles a LEFT JOIN categories c ON a.article_id = c.article_id WHERE a.article_id = ?;',
-		[id],
-	);
+	try {
+		const [article] = await connection.execute(
+			'SELECT a.article_id, a.title, a.content, c.category FROM articles a LEFT JOIN categories c ON a.article_id = c.article_id WHERE a.article_id = ?;',
+			[id],
+		);
 
-	if (article.length === 0) throw new ClientError('Invalid ID');
+		if (article.length === 0) throw new ClientError('Invalid ID');
 
-	return article;
+		return { code: 200, response: article };
+	} catch (error) {
+		if (error.statusCode === 400)
+			throw {
+				code: error.statusCode,
+				response: error.message,
+			};
+		throw { completeError: error };
+	} finally {
+		connection.end();
+	}
 }
 
 export async function createModel({ input }) {
-	// validate input data
-	const result = validateInputArticle(input);
-
-	if (!result.success) throw new ClientError('data entered is incorrect');
-
-	const newArticle = {
-		id: randomUUID(),
-		...result,
-	};
-
-	// connect and insert in DB
 	const connection = await connectionDB();
+
 	try {
+		// validate input data
+		const result = validateInputArticle(input);
+
+		if (!result.success) throw new ClientError('your input data is not valid');
+
+		const newArticle = {
+			id: randomUUID(),
+			...result,
+		};
+
+		//Insert data articles
 		await connection.beginTransaction();
 
 		await connection.execute(
@@ -71,50 +95,72 @@ export async function createModel({ input }) {
 		await connection.commit();
 
 		return {
-			message: 'the article was correctly inserted in the database',
-			article_id: newArticle.id,
+			code: 201,
+			message: `the article was insert in database with id ${newArticle.id}`,
 		};
 	} catch (error) {
 		await connection.rollback();
-		if (error instanceof Error)
-			return { success: false, message: error.message, code: error.code };
+
+		if (error.statusCode === 400)
+			throw {
+				code: error.statusCode,
+				response: error.message,
+			};
+		throw { completeError: error };
+	} finally {
+		connection.end();
 	}
 }
 
 export async function updateModel({ id, input }) {
-	const result = partialValidateInputArticle(input);
-
-	if (!result.success) throw new ClientError('data entered is incorrect');
-
 	const connection = await connectionDB();
-
-	let query = 'UPDATE articles SET ';
-	let params = [];
-
-	if (result.data.title !== undefined) {
-		query += 'title = ?, ';
-		params.push(result.data.title);
-	}
-
-	if (result.data.content !== undefined) {
-		query += 'content = ?, ';
-		params.push(result.data.content);
-	}
-
-	query = query.slice(0, -2);
-
-	query += ' WHERE article_id = ?';
-	params.push(id);
-
 	try {
+		//Validate input data
+		const result = partialValidateInputArticle(input);
+
+		if (!result.success) throw new ClientError('data entered is incorrect');
+
+		const [article] = await connection.execute(
+			'SELECT title FROM articles WHERE article_id = ?',
+			[id],
+		);
+
+		if (article.length === 0) throw new ClientError('invalid id');
+
+		//Building query to update the article
+
+		let query = 'UPDATE articles SET ';
+		let params = [];
+
+		if (result.data.title !== undefined) {
+			query += 'title = ?, ';
+			params.push(result.data.title);
+		}
+
+		if (result.data.content !== undefined) {
+			query += 'content = ?, ';
+			params.push(result.data.content);
+		}
+
+		query = query.slice(0, -2);
+
+		query += ' WHERE article_id = ?';
+		params.push(id);
+
 		await connection.execute(query, params);
 		return {
-			message: 'The article was correctly updated in the database',
-			article_id: id,
+			code: 204,
+			message: `The article with id ${id} was updated`,
 		};
 	} catch (error) {
-		if (error instanceof Error)
-			return { success: false, message: error.message, code: error.code };
+		if (error.statusCode === 400)
+			throw {
+				code: error.statusCode,
+				response: error.message,
+			};
+		throw { completeError: error };
+	} finally {
+		connection.end();
 	}
 }
 
@@ -122,13 +168,25 @@ export async function removeModel({ id }) {
 	const connection = await connectionDB();
 
 	try {
+		const [article] = await connection.execute(
+			'SELECT title FROM articles WHERE article_id = ?',
+			[id],
+		);
+
+		if (article.length === 0) throw new ClientError('invalid id');
+
 		await connection.execute('DELETE FROM articles WHERE article_id = ?', [id]);
 
 		return {
-			message: 'the article was correctly deleted',
-			article_id: id,
+			code: 204,
+			response: 'the article with id ${id} was correctly deleted',
 		};
 	} catch (error) {
-		return { success: false, message: error.message, code: error.code };
+		if (error.statusCode === 400)
+			throw {
+				code: error.statusCode,
+				response: error.message,
+			};
+		throw { completeError: error };
 	}
 }
